requirements.txt
python-dotenv
google-genai
google-cloud-aiplatform
pillow
grpcio==1.60.1